requirements.txt
pandas==1.4.3
gspread==5.4.0
python-dotenv==0.21.0
requests==2.28.1
